test(projects): isolate guide and policy bundle fixture owners
